refactor:list를 감싸는 wrapper 공통 컴포넌트 추가
