+++ src/refactoring/features/admin/product/product-list.tsx
@@ -1,13 +1,14 @@
 import { useProductStore } from "../../../store/product-store";
+import { ListWrapper } from "../../../ui/list-wrapper";
 import { ProductItem } from "./product-item";
 
 export const ProductList = () => {
   const { products } = useProductStore();
   return (
-    <div className="space-y-2">
+    <ListWrapper>
       {products.map((product, index) => (
         <ProductItem product={product} index={index} />
       ))}
-    </div>
+    </ListWrapper>
   );
 };
